+++ sdk/src/main/java/ly/count/android/sdk/ModuleContent.java
@@ -11,6 +11,7 @@ import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
+import org.jetbrains.annotations.NotNull;
 import org.json.JSONObject;
 
 public class ModuleContent extends ModuleBase {
@@ -19,10 +20,11 @@ public class ModuleContent extends ModuleBase {
     CountlyTimer countlyTimer;
     private boolean shouldFetchContents = false;
     private boolean isCurrentlyInContentZone = false;
-    private final int zoneTimerInterval;
+    private int zoneTimerInterval;
     private final ContentCallback globalContentCallback;
     private int waitForDelay = 0;
-    private final int CONTENT_START_DELAY_MS = 4000; // 4 seconds
+    int CONTENT_START_DELAY_MS = 4000; // 4 seconds
+    int REFRESH_CONTENT_ZONE_DELAY_MS = 2500; // 2.5 seconds
 
     ModuleContent(@NonNull Countly cly, @NonNull CountlyConfig config) {
         super(cly, config);
@@ -33,6 +35,27 @@ public class ModuleContent extends ModuleBase {
         countlyTimer = new CountlyTimer();
         zoneTimerInterval = config.content.zoneTimerInterval;
         globalContentCallback = config.content.globalContentCallback;
+    }
+
+    @Override
+    void onSdkConfigurationChanged(@NonNull CountlyConfig config) {
+        zoneTimerInterval = config.content.zoneTimerInterval;
+        if (!configProvider.getContentZoneEnabled()) {
+            exitContentZoneInternal();
+        } else {
+            if (!shouldFetchContents) {
+                exitContentZoneInternal();
+            }
+            waitForDelay = 0;
+            enterContentZoneInternal(null, 0);
+        }
+    }
+
+    @Override
+    void initFinished(@NotNull CountlyConfig config) {
+        if (configProvider.getContentZoneEnabled()) {
+            enterContentZoneInternal(null, 0);
+        }
     }
 
     void fetchContentsInternal(@NonNull String[] categories) {
@@ -77,22 +100,34 @@ public class ModuleContent extends ModuleBase {
         }, L);
     }
 
-    void registerForContentUpdates(@Nullable String[] categories, final int initialDelayMS) {
-        if (deviceIdProvider.isTemporaryIdEnabled()) {
-            L.w("[ModuleContent] registerForContentUpdates, temporary device ID is enabled, skipping");
+    private void enterContentZoneInternal(@Nullable String[] categories, final int initialDelayMS) {
+        if (!consentProvider.getConsent(Countly.CountlyFeatureNames.content)) {
+            L.w("[ModuleContent] enterContentZoneInternal, Consent is not granted, skipping");
             return;
         }
+
+        if (deviceIdProvider.isTemporaryIdEnabled()) {
+            L.w("[ModuleContent] enterContentZoneInternal, temporary device ID is enabled, skipping");
+            return;
+        }
+
+        if (isCurrentlyInContentZone) {
+            L.w("[ModuleContent] enterContentZoneInternal, already in content zone, skipping");
+            return;
+        }
+
+        shouldFetchContents = true;
 
         String[] validCategories;
 
         if (categories == null) {
-            L.w("[ModuleContent] registerForContentUpdates, categories is null, providing empty array");
+            L.w("[ModuleContent] enterContentZoneInternal, categories is null, providing empty array");
             validCategories = new String[] {};
         } else {
             validCategories = categories;
         }
 
-        L.d("[ModuleContent] registerForContentUpdates, categories: [" + Arrays.toString(validCategories) + "]");
+        L.d("[ModuleContent] enterContentZoneInternal, categories: [" + Arrays.toString(validCategories) + "]");
 
         int contentInitialDelay = initialDelayMS;
         long sdkStartTime = UtilsTime.currentTimestampMs() - Countly.applicationStart;
@@ -100,20 +135,21 @@ public class ModuleContent extends ModuleBase {
             contentInitialDelay += CONTENT_START_DELAY_MS;
         }
 
-        countlyTimer.startTimer(zoneTimerInterval, contentInitialDelay, () -> {
-            L.d("[ModuleContent] registerForContentUpdates, waitForDelay: [" + waitForDelay + "], shouldFetchContents: [" + shouldFetchContents + "], categories: [" + Arrays.toString(validCategories) + "]");
+        countlyTimer.startTimer(zoneTimerInterval, contentInitialDelay, new Runnable() {
+            @Override public void run() {
+                L.d("[ModuleContent] enterContentZoneInternal, waitForDelay: [" + waitForDelay + "], shouldFetchContents: [" + shouldFetchContents + "], categories: [" + Arrays.toString(validCategories) + "]");
+                if (waitForDelay > 0) {
+                    waitForDelay--;
+                    return;
+                }
 
-            if (waitForDelay > 0) {
-                waitForDelay--;
-                return;
+                if (!shouldFetchContents) {
+                    L.w("[ModuleContent] enterContentZoneInternal, shouldFetchContents is false, skipping");
+                    return;
+                }
+
+                fetchContentsInternal(validCategories);
             }
-
-            if (!shouldFetchContents) {
-                L.w("[ModuleContent] registerForContentUpdates, shouldFetchContents is false, skipping");
-                return;
-            }
-
-            fetchContentsInternal(validCategories);
         }, L);
     }
 
@@ -209,15 +245,6 @@ public class ModuleContent extends ModuleBase {
         }
     }
 
-    private void enterContentZoneInternal(@Nullable String[] categories, final int initialDelayMS) {
-        if (isCurrentlyInContentZone) {
-            L.w("[ModuleContent] enterContentZone, already in content zone, skipping");
-            return;
-        }
-        shouldFetchContents = true;
-        registerForContentUpdates(categories, initialDelayMS);
-    }
-
     private void exitContentZoneInternal() {
         shouldFetchContents = false;
         countlyTimer.stopTimer(L);
@@ -225,10 +252,9 @@ public class ModuleContent extends ModuleBase {
     }
 
     private void refreshContentZoneInternal() {
-        // Will be added with server config PR to not overcrowded this PR
-        // if (!configProvider.getRefreshContentZoneEnabled()) {
-        //    return;
-        //}
+        if (!configProvider.getRefreshContentZoneEnabled()) {
+            return;
+        }
 
         if (isCurrentlyInContentZone) {
             L.w("[ModuleContent] refreshContentZone, already in content zone, skipping");
@@ -241,7 +267,7 @@ public class ModuleContent extends ModuleBase {
 
         _cly.moduleRequestQueue.attemptToSendStoredRequestsInternal();
 
-        enterContentZoneInternal(null, 2500);
+        enterContentZoneInternal(null, REFRESH_CONTENT_ZONE_DELAY_MS);
     }
 
     public class Content {