+++ sdk/src/main/java/ly/count/android/sdk/ModuleDeviceId.java
@@ -81,7 +81,7 @@ public class ModuleDeviceId extends ModuleBase implements OpenUDIDProvider, Devi
         deviceIdInstance.changeToCustomId(deviceId);
 
         // trigger fetching if the temp id given on init
-        _cly.moduleConfiguration.fetchConfigFromServer();
+        _cly.moduleConfiguration.fetchConfigFromServer(_cly.config_);
 
         //update stored request for ID change to use this new ID
         replaceTempIDWithRealIDinRQ(deviceId);